+++ src/commands/build.ts
@@ -1,3 +1,4 @@
+import * as fs from 'fs';
 import { ExtensionContext, Uri, window } from 'vscode';
 import { Commands, BaseCommand } from '.';
 import StreamsBuild from '../build';
@@ -33,15 +34,30 @@ export default class BuildCommand implements BaseCommand {
         switch (this.commandName) {
             case Commands.BUILD.APP_DOWNLOAD:
             case Commands.BUILD.APP_SUBMIT:
+            case Commands.BUILD.APP_IMAGE:
                 return StreamsBuild.buildApp(filePaths[0], this._buildType)
                     .catch((error) => { throw error; });
             case Commands.BUILD.MAKE_DOWNLOAD:
             case Commands.BUILD.MAKE_SUBMIT:
+            case Commands.BUILD.MAKE_IMAGE:
                 return StreamsBuild.buildMake(filePaths[0], this._buildType)
                     .catch((error) => { throw error; });
             case Commands.BUILD.SUBMIT:
                 return StreamsBuild.submit(filePaths)
                     .catch((error) => { throw error; });
+            case Commands.BUILD.IMAGE:
+                return StreamsBuild.buildImage(filePaths)
+                    .catch((error) => { throw error; });
+            case Commands.ENVIRONMENT.ADD_TOOLKIT_TO_BUILD_SERVICE:
+                let selected = null;
+                if (filePaths && filePaths.length && fs.existsSync(filePaths[0])) {
+                    selected = filePaths[0];
+                }
+                return StreamsBuild.addToolkitToBuildService(selected)
+                    .catch((error) => { throw error; })
+            case Commands.ENVIRONMENT.REMOVE_TOOLKITS_FROM_BUILD_SERVICE:
+                return StreamsBuild.removeToolkitsFromBuildService()
+                    .catch((error) => { throw error; })
             default:
                 return null;
         }