+++ src/utils/vscode.ts
@@ -1,7 +1,18 @@
-import { window } from 'vscode';
+import * as path from 'path';
+import {
+    env, Uri, window, workspace, WorkspaceFolder
+} from 'vscode';
 
 const YES_LABEL = 'Yes';
 const NO_LABEL = 'No';
+
+/**
+ * Copy text to the user's clipboard
+ * @param value    The value to copy to the clipboard
+ */
+async function copyToClipboard(value: string): Promise<void> {
+    await env.clipboard.writeText(value);
+}
 
 /**
  * Show a confirmation dialog
@@ -20,8 +31,65 @@ async function showConfirmationDialog(label: string, yesCallbackFn: Function): P
 }
 
 /**
+ * Get the workspace folder paths
+ */
+function getWorkspaceFolderPaths(): string[] {
+    return workspace.workspaceFolders
+        ? workspace.workspaceFolders.map(
+            (folder: WorkspaceFolder) => folder.uri.fsPath
+        )
+        : [];
+}
+
+/**
+ * Determine whether a folder or a folder above it is a workspace folder
+ * @param folderPath the folder path to check
+ */
+function isWorkspaceFolder(folderPath: string): boolean {
+    const rootFolderPath = path.parse(process.cwd()).root;
+    const workspaceFolderPaths = getWorkspaceFolderPaths();
+    const notWorkspaceFolder = (folderPath: string): boolean =>
+        !workspaceFolderPaths.includes(folderPath);
+    const notRootFolder = (folderPath: string): boolean =>
+        folderPath !== rootFolderPath;
+    // Traverse upwards from starting folder until we reach a workspace folder or the root folder
+    let currentFolderPath = folderPath;
+    while (
+        notWorkspaceFolder(currentFolderPath) &&
+        notRootFolder(currentFolderPath)
+    ) {
+        currentFolderPath = path.dirname(currentFolderPath);
+    }
+    // If we are not at the root folder, then it is a workspace folder
+    return currentFolderPath !== rootFolderPath;
+}
+
+/**
+ * Add one or more folders to the workspace
+ * @param folderPaths the folder paths to add
+ */
+function addFoldersToWorkspace(folderPaths: string[]): void {
+    const foldersToAdd = folderPaths
+        .filter((folderPath) => !isWorkspaceFolder(folderPath))
+        .map((folderPath) => ({ uri: Uri.file(folderPath) }));
+    if (foldersToAdd.length) {
+        workspace.updateWorkspaceFolders(
+            workspace.workspaceFolders ? workspace.workspaceFolders.length : 0,
+            null,
+            ...foldersToAdd
+        );
+    }
+}
+
+/**
  * VS Code utilities
  */
-const VSCode = { showConfirmationDialog };
+const VSCode = {
+    addFoldersToWorkspace,
+    copyToClipboard,
+    getWorkspaceFolderPaths,
+    isWorkspaceFolder,
+    showConfirmationDialog
+};
 
 export default VSCode;